arreglar valores nulos

+++ relationaldb/src/main/java/com/finproyectodam/relationaldb/repository/UsuariosRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuariosRepository extends JpaRepository<Usuario, Long> {
+public interface UsuariosRepository extends JpaRepository<Usuario, Integer> {
 
 }
